feat: COINjecture Network B v5 - NP-hard consensus blockchain
Major fix: libp2p connections on Linux
- Replaced LocalSet/spawn_local with multi-threaded tokio runtime
- Fixes Noise/Yamux handshake stalls on Linux droplets

Features:
- 6 specialized node types (Light, Full, Archive, Validator, Bounty, Oracle)
- NP-hard problem consensus
- Gossipsub P2P networking
- HuggingFace dataset sync
- Privacy marketplace

+++ tokenomics/src/emission.rs
@@ -271,7 +271,7 @@ impl EmissionCalculator {
         let remaining = target_supply.saturating_sub(self.metrics.circulating_supply);
         
         if current_emission > 0 {
-            Some(remaining / current_emission)
+            Some((remaining / current_emission as u128) as u64)
         } else {
             None
         }
